requirements.txt
anthropic==0.49.0
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.28.0
gspread==6.2.1
google-auth==2.48.0
google-auth-oauthlib==1.2.4
